Nest ten Lira satellites under the Lira namespace
`LiraAdvisory`, `LiraDelta`, `LiraError`, `LiraHash`, `LiraManifest`,
`LiraPayload`, `LiraRealm`, `LiraSchemas`, `LiraTree` and `LiraValidators` become
members of `object Lira`.

`LiraAssembler` (reliquary.derive) and `LiraBundle` (anthology.lira) stay: both
are downstream of reliquary.core, so R6 blocks them.

`Lira.Hash.Domain` has a case named `Manifest`, which the newly-nested
`Lira.Manifest` made ambiguous inside the enum's own body — the sibling-level
clash that also bit `Wit.Function`. The pattern is qualified as `Domain.Manifest`.
