--- lib/reliquary/src/core/reliquary.LiraPayload.scala
@@ -1,66 +1,0 @@
-                                                                                                  /*
-┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
-┃                                                                                                  ┃
-┃                                                   ╭───╮                                          ┃
-┃                                                   │   │                                          ┃
-┃                                                   │   │                                          ┃
-┃   ╭───────╮╭─────────╮╭───╮ ╭───╮╭───╮╌────╮╭────╌┤   │╭───╮╌────╮╭────────╮╭───────╮╭───────╮   ┃
-┃   │   ╭───╯│   ╭─╮   ││   │ │   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮  ││   ╭───╯│   ╭───╯   ┃
-┃   │   ╰───╮│   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╰─╯  ││   ╰───╮│   ╰───╮   ┃
-┃   ╰───╮   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╭────╯╰───╮   │╰───╮   │   ┃
-┃   ╭───╯   ││   ╰─╯   ││   ╰─╯   ││   │ │   ││   ╰─╯   ││   │ │   ││   ╰────╮╭───╯   │╭───╯   │   ┃
-┃   ╰───────╯╰─────────╯╰────╌╰───╯╰───╯ ╰───╯╰────╌╰───╯╰───╯ ╰───╯╰────────╯╰───────╯╰───────╯   ┃
-┃                                                                                                  ┃
-┃    Soundness, version 0.64.0.                                                                    ┃
-┃    © Copyright 2021-25 Jon Pretty, Propensive OÜ.                                                ┃
-┃                                                                                                  ┃
-┃    The primary distribution site is:                                                             ┃
-┃                                                                                                  ┃
-┃        https://soundness.dev/                                                                    ┃
-┃                                                                                                  ┃
-┃    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file     ┃
-┃    except in compliance with the License. You may obtain a copy of the License at                ┃
-┃                                                                                                  ┃
-┃        https://www.apache.org/licenses/LICENSE-2.0                                               ┃
-┃                                                                                                  ┃
-┃    Unless required by applicable law or agreed to in writing,  software distributed under the    ┃
-┃    License is distributed on an "AS IS" BASIS,  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,    ┃
-┃    either express or implied. See the License for the specific language governing permissions    ┃
-┃    and limitations under the License.                                                            ┃
-┃                                                                                                  ┃
-┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
-                                                                                                  */
-package reliquary
-
-import anticipation.*
-import contingency.*
-import gossamer.*
-import pneumatic.*
-import zephyrine.*
-
-import LiraError.Reason
-
-// The compression envelope (§8.1) around the blob stream. The compressed bytes are not part of
-// any identity: `payload.hash` is the blob-domain hash of the *decompressed* stream (§8.4), so
-// two files differing only in compressor output carry equal implementation identities. The
-// declared decompressed length is enforced as a hard limit (L102) and the declared hash is
-// verified (L105).
-//
-// pneumatic's Brotli engine materializes the whole stream to decode it, so the length cap is
-// currently enforced on the materialized result; when a streaming Brotli decoder lands, this is
-// the seam at which decompression should abort mid-stream instead.
-object LiraPayload:
-
-  def compress(blobStream: Data): Data = blobStream.compress[Brotli]
-
-  def hash(blobStream: Data): Data = LiraHash(LiraHash.Domain.Blob, blobStream)
-
-  def decompress(compressed: Data, length: Long, declaredHash: Data): Data raises LiraError =
-    val result =
-      try compressed.decompress[Brotli] catch case error: Exception =>
-        abort(LiraError(Reason.MalformedPayload(t"the payload does not decompress")))
-
-    if result.length.toLong != length then abort(LiraError(Reason.PayloadLength(length)))
-    if Blob.compare(hash(result), declaredHash) != 0 then abort(LiraError(Reason.PayloadHash))
-
-    result
