+++ lib/reliquary/src/core/soundness_reliquary_core.scala
@@ -34,7 +34,5 @@ package soundness
 
 export reliquary.{Atom, AtomClass, Atomization, AtomReference, AtomsBlob, Blob, BlobStream,
     Blobstore, Buildpath, CapabilityDiscipline, Discipline, DisciplineError, EcosystemProfile,
-    Grade, Lineage, Lira, LiraAdvisory,
-    LiraDelta, LiraError, LiraHash, LiraManifest, LiraPayload, LiraSchemas, LiraTree,
-    LiraValidators, LiraRealm, ManifestSigning, OpaqueDiscipline, Overlay, Publication,
+    Grade, Lineage, Lira, ManifestSigning, OpaqueDiscipline, Overlay, Publication,
     Replacement, Section, Snapshot, TreeEntry, TreePath, UsesBlob, Verification, Versioning}
